Not required to install.. just download and run file
