--- installation/Pipfile
@@ -1,13 +1,0 @@
-[[source]]
-url = "https://pypi.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[dev-packages]
-
-[packages]
-tkinter 
-re
-
-[requires]
-python_version = "3.6"
